fix MCP dependency visibility reporting
